requirements.txt
beautifulsoup4==4.13.5
lxml==6.0.2
psycopg2==2.9.10
pytest==9.0.1
pytest-mock==3.15.1
python-dotenv==1.2.1
requests==2.32.5